Remove extension to getCapabilties() in favor of MediaCapabilities
Rerefences to scalabilityModes in getCapabilties('video').codecs have
been removed and replaced with references to the MediaCapabilities
API when appropriate and an example for the MediaCapabilites API
usage have been added.

The SFM capabilities description has not been updated since it is not
normative or reflective of the Javascript APIs we support anyway and
falls into the domain of native APIs.